split validation
